+++ src/Transformer/ObjectToObjectTransformer.php
@@ -116,8 +116,11 @@ final class ObjectToObjectTransformer implements TransformerInterface, MainTrans
 
         // save object to cache
 
-        $context(ObjectCache::class)
-            ->saveTarget($source, $targetType, $target, $context);
+        $context(ObjectCache::class)->saveTarget(
+            source: $source,
+            targetType: $targetType,
+            target: $target,
+        );
 
         // map properties