+++ src/Transformer/TraversableToTraversableTransformer.php
@@ -89,8 +89,11 @@ final class TraversableToTraversableTransformer implements TransformerInterface,
 
         // Add to cache
 
-        $context(ObjectCache::class)
-            ->saveTarget($source, $targetType, $target, $context);
+        $context(ObjectCache::class)->saveTarget(
+            source: $source,
+            targetType: $targetType,
+            target: $target,
+        );
 
         return $target;
     }